requirements-dev.txt
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
